[refactor] Update package structure to use 'models' for domain classes
