--- domain/src/main/java/com/adrifernandev/marsroverchallenge/domain/model/Direction.kt
@@ -1,5 +1,0 @@
-package com.adrifernandev.marsroverchallenge.domain.model
-
-enum class Direction {
-    N, E, S, W
-}
