--- domain/src/main/java/com/adrifernandev/marsroverchallenge/domain/model/Position.kt
@@ -1,6 +1,0 @@
-package com.adrifernandev.marsroverchallenge.domain.model
-
-data class Position(
-    val x: Int,
-    val y: Int
-)
