+++ domain/src/main/java/com/adrifernandev/marsroverchallenge/domain/models/Direction.kt
@@ -1,0 +1,5 @@
+package com.adrifernandev.marsroverchallenge.domain.models
+
+enum class Direction {
+    N, E, S, W
+}
