+++ domain/src/main/java/com/adrifernandev/marsroverchallenge/domain/models/Plateau.kt
@@ -1,4 +1,4 @@
-package com.adrifernandev.marsroverchallenge.domain.model
+package com.adrifernandev.marsroverchallenge.domain.models
 
 data class Plateau(
     val topRightCornerPosition: Position
