+++ domain/src/main/java/com/adrifernandev/marsroverchallenge/domain/models/Position.kt
@@ -1,0 +1,6 @@
+package com.adrifernandev.marsroverchallenge.domain.models
+
+data class Position(
+    val x: Int,
+    val y: Int
+)
